api/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cssselect2==0.6.0
Django==4.1
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.24.2
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.1
reportlab==3.6.11
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tinycss2==1.1.1
Unidecode==1.1.2
uritemplate==4.1.1
urllib3==1.26.12
webencodings==0.5.1
